refactor(social-messenger-ts): use default SDK ChannelList
