+++ social-messenger-ts/src/components/index.ts
@@ -2,8 +2,7 @@ export { default as AvatarGroup } from './AvatarGroup/AvatarGroup';
 export { default as ChannelInner } from './ChannelInner/ChannelInner';
 export { default as CreateChannel } from './CreateChannel/CreateChannel';
 export { default as MessagingChannelHeader } from './MessagingChannelHeader/MessagingChannelHeader';
-export { default as MessagingChannelList } from './MessagingChannelList/MessagingChannelList';
-export { default as MessagingChannelListHeader } from './MessagingChannelList/MessagingChannelListHeader';
+export { default as MessagingChannelListHeader } from './MessagingChannelListHeader/MessagingChannelListHeader';
 export { default as MessagingChannelPreview } from './MessagingChannelPreview/MessagingChannelPreview';
 export { default as MessagingSidebar } from './MessagingSidebar/MessagingSidebar';
 export { default as MessagingThreadHeader } from './MessagingThreadHeader/MessagingThreadHeader';
