+++ social-messenger-ts/src/styles/overrides/MessagingSidebar/MessagingSidebar.layout.css
@@ -4,9 +4,6 @@
     grid-template-rows: min-content;
 }
 
-.str-chat__channel-list {
-    padding: 20px 10px;
-}
 
 .messaging__channel-list__header {
     display: flex;
